feat(web): prefetch assets for asset picker
