+++ libs/web/src/components/common/AssetDataPrefetcher.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import {usePrefetchAssetData} from "@/src/hooks";
+
+export const AssetDataPrefetcher = () => {
+  usePrefetchAssetData();
+  return null;
+};
