+++ libs/web/src/hooks/index.ts
@@ -54,3 +54,4 @@ export {useWalletTransactions} from "./useWalletTransactions";
 export {usePreviewAddLiquidity} from "./usePreviewAddLiquidity";
 export {useDocumentTitle} from "./useDocumentTitle";
 export {useProtocolStats} from "./useProtocolStats";
+export {usePrefetchAssetData} from "./usePrefetchAssetData";
